feat: Refactor product option and variation classes
This commit refactors the product option and variation classes to improve code organization and maintainability. The `ProductOption` class has been moved to the `variationOption` package, and the `ProductVariation` class has been renamed to `Variation` and moved to the `variation` package. The necessary changes have also been made to the import statements and package declarations in the affected files.